response/requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Limiter
python-dotenv
requests
PyMySQL
Werkzeug
bleach